certbot: Migrate to renewal-hooks/deploy for 0.19
certbot now provides a built in '/etc/letsencrypt/renewal-hooks'
directory.  Use this instead of a custom post-hook.

Also, use the 'deploy' hook.  This functions similar to the renewal
hook, except it's also run on first-time setup.  Renewal hooks are
run once-per-certificate, but as there's only one certificate, this
shouldn't matter.

The call to certbot_reload in certbot-setup.sh "configure" could be
removed, but the testing environment could not be used to verify
this.