web: Protect against HTTP Host header caching
Add a check for expected hostnames when redirecting from HTTP to
HTTPS.  Otherwise, it's possible for an attacker to send a request
with a Host header to another domain, get the HTTP redirect cached by
a proxy, and redirect any victims behind this proxy.

This shouldn't happen, but it's a large enough of a concern that
Let's Encrypt changed their setup, so being safe here's a good idea,
too.

See https://community.letsencrypt.org/t/security-issue-with-redirects-added-by-certbots-nginx-plugin/51493
And https://bjornjohansen.no/redirect-to-https-with-nginx