quassel-search: Switch to 3.0 branch
Switch to 3.0 branch by default, replacing version 2.0.  This also
changes the configuration format, so any deployed pillar files must
upgrade to a 3.0 branch as well.

Update qrs_config.php to the new-style format.  Specify a timeout of
5 seconds by default; local machine shouldn't have that many issues.
This may be exposed as a pillar configuration option later.

Pretty much all features from 2.0 are in 3.0, it's faster, shinier,
etc :)