quassel: Enable logging to syslog
Enable Quassel logging to syslog, not just the log file.  This makes
it easier to see the logs by putting them in one central place and
allowing systemd to manage the output.

In the future, the log file may be removed, relying on syslog only.

Fix Quassel core to restart on configuration change, too.