status: Add optional system monitoring to remote
Add Telegraf statistics gathering daemon, disabled by default.  If
enabled and configured, it will gather system statistics and uptime
stats, sending it to a remote HTTP Listener on another Telegraf
server (proxied through nginx or however).

This is disabled by default, and only reports to the server you
specify in the pillar data.  No third-party is involved!

Add a bunch of configuration files to support this.

On the remote server, you might setup Telegraf, InfluxDB, and
Grafana.

This configuration goes well with these Grafana dashboards:
> https://grafana.com/dashboards/2690 "Internet Uptime Monitor"
> https://grafana.com/dashboards/3967 "System"

And a custom variant of Internet Uptime Monitor that checks HTTP
instead of ping (to be posted eventually).