backup: Fix GZIP environment variable deprecation
Fix reliance on the "GZIP" environment variable to set gzip to use
rsyncable compression.  The proper method involves setting tar's
--use-compress-program option, which is supported all the way back to
Ubuntu 16.04 LTS.

This does not impact salt-box-quassel's current backup scripts as
Quassel does not backup any data directories, but it is included for
correctness and consistency with other deployments.

See https://stackoverflow.com/questions/46167772/with-gnu-gzip-environment-variable-deprecated-how-to-control-zlib-compression-v